backend/requirements.txt
astroid==2.2.5
cffi==1.15.1
Click==7.0
cryptography==37.0.2
ecdsa==0.13.2
Flask==1.0.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.5.0
future==0.17.1
greenlet==1.1.2
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.0
MarkupSafe==1.1.1
mccabe==0.6.1
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.3.1
pylint==2.3.1
python-jose==3.3.0
python-jose-cryptodome==1.3.2
rsa==4.8
six==1.12.0
SQLAlchemy==1.4.39
typed-ast==1.4.2
Werkzeug==0.15.6
wrapt==1.11.1
